+++ axelor-crm/src/main/java/com/axelor/apps/crm/service/batch/BatchEventReminder.java
@@ -33,7 +33,7 @@ import com.axelor.apps.message.service.MailAccountService;
 import com.axelor.apps.message.service.MessageService;
 import com.axelor.db.JPA;
 import com.axelor.exception.AxelorException;
-import com.axelor.exception.db.IException;
+import com.axelor.exception.db.repo.ExceptionOriginRepository;
 import com.axelor.exception.db.repo.TraceBackRepository;
 import com.axelor.exception.service.TraceBackService;
 import com.axelor.i18n.I18n;
@@ -104,7 +104,7 @@ public class BatchEventReminder extends BatchStrategy {
                       I18n.get(IExceptionMessage.BATCH_EVENT_REMINDER_1),
                       eventReminderRepo.find(eventReminder.getId()).getEvent().getSubject()),
                   e),
-              IException.CRM,
+              ExceptionOriginRepository.CRM,
               batch.getId());
 
           incrementAnomaly();
@@ -229,7 +229,7 @@ public class BatchEventReminder extends BatchStrategy {
                       I18n.get("Event") + " %s",
                       eventRepo.find(eventReminder.getEvent().getId()).getSubject()),
                   e),
-              IException.CRM,
+              ExceptionOriginRepository.CRM,
               batch.getId());
 
           incrementAnomaly();