+++ CHANGELOG.md
@@ -179,6 +179,7 @@
 - ACCOUNTING : year and accountSet fields are set empty when company is changed
 - ACCOUNTING CONFIGURATION AND REPORTED BALANCE : add a new journal parameter in accounting configuration named reported balance journal and this new parameter defines the journal used in the reported balance move line creation.
 - PARTNER: Checks the consistency of the address list of a partner.
+- EXCEPTION : Removing deprecated interface IException
 - PORTAL CLIENT : Add config to choose the type of connection (from ABS or external)
 - DEMO DATA : Rename field data.
 - SALE ORDER/ PURCHASE ORDER : Removal of spacer.